+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/doorlock/UnlockDoorResponse.java
@@ -102,7 +102,7 @@ public class UnlockDoorResponse extends ZclDoorLockCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        status = (Integer) deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
+        status = deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
     }
 
     @Override
