+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/greenpower/GpSinkCommissioningMode.java
@@ -185,10 +185,10 @@ public class GpSinkCommissioningMode extends ZclGreenPowerCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        options = (Integer) deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
-        gpmAddrForSecurity = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
-        gpmAddrForPairing = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
-        sinkEndpoint = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
+        options = deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
+        gpmAddrForSecurity = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        gpmAddrForPairing = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        sinkEndpoint = deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
     }
 
     @Override
