+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/rssilocation/SetDeviceConfigurationCommand.java
@@ -211,11 +211,11 @@ public class SetDeviceConfigurationCommand extends ZclRssiLocationCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        power = (Integer) deserializer.deserialize(ZclDataType.SIGNED_16_BIT_INTEGER);
-        pathLossExponent = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
-        calculationPeriod = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
-        numberRssiMeasurements = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
-        reportingPeriod = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        power = deserializer.deserialize(ZclDataType.SIGNED_16_BIT_INTEGER);
+        pathLossExponent = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        calculationPeriod = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        numberRssiMeasurements = deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
+        reportingPeriod = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
     }
 
     @Override
