+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/thermostat/GetWeeklyScheduleResponse.java
@@ -240,12 +240,12 @@ public class GetWeeklyScheduleResponse extends ZclThermostatCommand {
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        numberOfTransitions = (Integer) deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
-        dayOfWeek = (Integer) deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
-        mode = (Integer) deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
-        transition = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
-        heatSet = (Integer) deserializer.deserialize(ZclDataType.SIGNED_16_BIT_INTEGER);
-        coolSet = (Integer) deserializer.deserialize(ZclDataType.SIGNED_16_BIT_INTEGER);
+        numberOfTransitions = deserializer.deserialize(ZclDataType.ENUMERATION_8_BIT);
+        dayOfWeek = deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
+        mode = deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
+        transition = deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER);
+        heatSet = deserializer.deserialize(ZclDataType.SIGNED_16_BIT_INTEGER);
+        coolSet = deserializer.deserialize(ZclDataType.SIGNED_16_BIT_INTEGER);
     }
 
     @Override
